Moved some classes out of geometry.timeseries for clarity
